Proto-Build 3
This is Week 3 of working on my Intermediate Game Development Midterm.

The game is actually almost done! It seems as though every midterm game I work on, I finish it must faster than I expected. Is this a good sign...?!

Well, truthfully, this isn't done yet. The visuals are not all complete, and only have the essentials. I still need to model the needed objects, but that didn't take long last time. Music is something I am going to learn in class on Tuesday, March 12, so hopefully that answers all my questions. Plus, some bugs could be squashed as well.....



Wish me luck!