Housecleaning: Cleaner Folders/Scripts, More Models, Music and Animation
This time, this is a big Housecleaning update that helps organize my workspace as well as juice up the start and middle of the game. This will all help me in determining how the ending should be like, and what I want to do with it. (Spiral staircase made out of the cleaning objects, perhaps?)

Also, BECAREFUL WITH CHANGING OBJECT 3, FOR SOME REASON IT FREAKS OUT